fix errors when running the preview server on the demo
This was mostly caused because `pnpm` doesn't allow for resolving
`@react-email/render` when only `@react-email/components` is installed
like npm or other package managers do.

The solution for this was overriding the resolution of the module to get
`renderAsync` from, and then trying both `@react-email/render` and
`@react-email/components` until it actually errors.